+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2.4 (in-progress)
 ================
+* [#140](https://github.com/dblock/oshi/pull/140): Added process and thread counts - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here.
 
 2.3 (4/14/2016)